requirements.txt
termcolor
python-slugify
zmq
cement
PyYAML
docker-py
dockerpty